intl rankings

+++ src/types/rankings.rs
@@ -3,9 +3,9 @@ use serde::Deserialize;
 #[derive(Debug, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Rankings {
-    pub ranking: u32,
+    pub ranking: f64,
     pub name: String,
     pub weight_class: String,
     pub percent_a: f64,
-    pub total: i32,
+    pub total: f64,
 }
